requirements.txt
# Flask and related packages
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3

# Database
SQLAlchemy==2.0.19
alembic==1.11.1

# Web scraping and API requests
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
uuid==1.30

# Data validation and processing
deepeval==0.9.1
python-dateutil==2.8.2
pytz==2023.3

# Production deployment
gunicorn==21.2.0
psycopg2-binary==2.9.6  # For PostgreSQL on Heroku

# Testing and development
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.7.0